ApplicationsFAQ Refactoring
