+++ src/pages/organizers.js
@@ -5,9 +5,19 @@ import Footer from '../components/Footer'
 import ApplicationsIntro from '../components/ApplicationsIntro'
 import ApplicationsProcess from '../components/ApplicationsProcess'
 import ApplicationsCommittees from '../components/ApplicationsCommittees'
-import ApplicationsFAQs from '../components/ApplicationsFAQs'
 import ClosedImg from '../images/applications-closed.svg'
 import OpenImg from '../images/applications-open.svg'
+import FAQs from '../components/FAQs'
+
+const questions = {
+  'Who can be an organizer for PennApps?':
+          'Only Penn students can be organizers. If you’re a current Penn student, look out for an application in April if you are interested in joining us!',
+  'Do I need previous experience to be an organizer?':
+          'No previous experience required! We are looking for people who are passionate about creating an exciting and inclusive space to meet others and learn about tech and entrepreneurship and are able to commit to the time to be an organizer.',
+  'What is the time commitment?':
+          'Expect to commit 2 - 4 hours per week to PennApps meetings and tasks.',
+
+}
 
 export function Head() {
   return (
@@ -47,7 +57,7 @@ export default function Organizers() {
         <ApplicationsIntro />
         <ApplicationsProcess />
         <ApplicationsCommittees />
-        <ApplicationsFAQs />
+        <FAQs content={questions}/>
       </div>
       <Footer />
     </>
